Improve app record detail

+++ src/test/wmbus_selftest_support.c
@@ -219,13 +219,13 @@ void wmbus_selftest_describe_first_record(
     snprintf(
         out,
         out_size,
-        "records=%u q=%u vt=%u scale=%d value=%llu storage=%u data_len=%u",
+        "records=%u q=%u vt=%u scale=%d value=%llu storage=%llu data_len=%u",
         (unsigned int)packet->application.record_count,
         (unsigned int)rec->quantity,
         (unsigned int)rec->value_type,
         (int)rec->scale10,
         (unsigned long long)rec->value_unsigned,
-        (unsigned int)rec->storage_no,
+        (unsigned long long)rec->storage_no,
         (unsigned int)rec->data_len);
 }
